backend/requirements.txt
Flask
python-dotenv
Werkzeug
Flask
Flask-JWT-Extended
Flask-Cors
pymongo
python-dotenv
dnspython
Werkzeug
